requirements.txt
Flask==0.12.2
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2